+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 - use environment variable to define secret_token and secret_key_base
 - complete refactor of Heroku app creation and configuration
+- ask for user credentials before creating the Heroku app
 
 ### bug fixes